#13 Rename to cryptography_services.py to avoid collision with cryptography library, remove Encryption class as result of over-engineering creep
